demo: register a custom sparkle icon on the root TerrenoProvider
Register one custom SVG icon (sparkle) via the demo root TerrenoProvider's
icons prop so it can be exercised app-wide, and simplify the Icon CustomIcons
story to use the globally-registered icon by name.
