+++ demo/app/_layout.tsx
@@ -1,3 +1,4 @@
+import {customIcons} from "@components";
 import {TerrenoProvider} from "@terreno/ui";
 import {Slot} from "expo-router";
 import {GestureHandlerRootView} from "react-native-gesture-handler";
@@ -6,7 +7,7 @@ const RootLayout = () => {
   // TODO: Store dev/demo in AsyncStorage to persist.
   return (
     <GestureHandlerRootView style={{flex: 1}}>
-      <TerrenoProvider>
+      <TerrenoProvider icons={customIcons}>
         <Slot initialRouteName={process.env.NODE_ENV === "development" ? "dev" : "demo"} />
       </TerrenoProvider>
     </GestureHandlerRootView>
