+++ demo/components/index.tsx
@@ -1,3 +1,4 @@
+export * from "./customIcons";
 export * from "./DemoHomePage";
 export * from "./DevHomePage";
 export * from "./ErrorBoundary";
